Add files via upload
Added files necessary to run the MyArm300pi_EE_test_pt3.usd out of the box, not tested that works but it should work If you download the robot_data file to C:/Omniverse_Files. Wiki pending. Included are scripts to use the gripper. This should also be usable in VR in the omni only architecture.